Pull some modules out in Minimax
